add search modal
